[#513] Address pr feedback
