+++ src/electron/electron/main/services/utils/helpers.ts
@@ -24,3 +24,19 @@ export function generateAlphaNumericString(length: number = 0): string {
   }
   return result;
 }
+
+/**
+ * Formats a date as a SQLite-local datetime string.
+ *
+ * Queries that compare against `datetime(..., 'localtime')` need local wall-clock parameters
+ * instead of UTC ISO strings.
+ */
+export function formatSqliteLocalDateTime(date: Date): string {
+  const year = date.getFullYear();
+  const month = `${date.getMonth() + 1}`.padStart(2, '0');
+  const day = `${date.getDate()}`.padStart(2, '0');
+  const hours = `${date.getHours()}`.padStart(2, '0');
+  const minutes = `${date.getMinutes()}`.padStart(2, '0');
+  const seconds = `${date.getSeconds()}`.padStart(2, '0');
+  return `${year}-${month}-${day} ${hours}:${minutes}:${seconds}`;
+}
